# docs/openapi/sla_chat.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SLA Chat Service",
    "description": "LLM gateway for SLA and BI context.",
    "version": "1.0.0"
  },
  "paths": {
    "/healthz": {
      "get": {
        "summary": "Healthcheck",
        "operationId": "healthcheck_healthz_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/v1/sla/chat": {
      "post": {
        "summary": "Chat",
        "operationId": "chat_v1_sla_chat_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ChatRequest": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "question": {
            "type": "string",
            "title": "Question"
          },
          "top_k": {
            "type": "integer",
            "maximum": 10.0,
            "minimum": 1.0,
            "title": "Top K",
            "default": 5
          },
          "customer_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Customer Id",
            "description": "Optional customer identifier for downstream filtering."
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "question"
        ],
        "title": "ChatRequest"
      },
      "ChatResponse": {
        "properties": {
          "answer": {
            "type": "string",
            "title": "Answer"
          },
          "model": {
            "type": "string",
            "title": "Model"
          },
          "references": {
            "items": {
              "$ref": "#/components/schemas/ContextReference"
            },
            "type": "array",
            "title": "References"
          },
          "usage": {
            "additionalProperties": true,
            "type": "object",
            "title": "Usage"
          }
        },
        "type": "object",
        "required": [
          "answer",
          "model",
          "references",
          "usage"
        ],
        "title": "ChatResponse"
      },
      "ContextReference": {
        "properties": {
          "entry_id": {
            "type": "string",
            "title": "Entry Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "score": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Score"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata"
          }
        },
        "type": "object",
        "required": [
          "entry_id",
          "title",
          "score",
          "content",
          "metadata"
        ],
        "title": "ContextReference"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Mind-Q SLA Chat Service"
    }
  ]
}